requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.10.0
sqlalchemy>=2.0.35
openai>=1.51.0
python-multipart>=0.0.12
python-dotenv>=1.0.1
python-dateutil>=2.9.0
httpx>=0.27.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
aiosqlite>=0.20.0
psycopg2-binary>=2.9.9
